+++ R/int_utils.R
@@ -354,18 +354,19 @@ auto_mod_vals <-
 ## Centering
 
 center_ss <- function(d, weights, facvars = NULL, fvars, pred, resp, modx,
-                        survey, design = NULL, mod2, wname, offname, centered) {
+                      survey, design = NULL, mod2, wname, offname, centered,
+                      at = NULL) {
 
   # Just need to pick a helper function based on survey vs no survey
   if (survey == TRUE) {
 
     out <- center_ss_survey(d, weights, facvars, fvars, pred, resp, modx,
-                              survey, design, mod2, wname, offname, centered)
+                            survey, design, mod2, wname, offname, centered, at)
 
   } else {
 
     out <- center_ss_non_survey(d, weights, facvars, fvars, pred, resp, modx,
-                                  mod2, wname, offname, centered)
+                                mod2, wname, offname, centered, at)
 
   }
 
@@ -377,7 +378,7 @@ center_ss <- function(d, weights, facvars = NULL, fvars, pred, resp, modx,
 ## If not svydesign, centering is fairly straightforward
 
 center_ss_non_survey <- function(d, weights, facvars = NULL, fvars, pred,
-                                   resp, modx, mod2, wname, offname, centered) {
+                                 resp, modx, mod2, wname, offname, centered, at) {
 
   omitvars <- c(pred, resp, modx, mod2, wname, offname)
 
@@ -389,8 +390,8 @@ center_ss_non_survey <- function(d, weights, facvars = NULL, fvars, pred,
   if (centered[1] != "all" && centered[1] != "none") {
 
     if (any(omitvars %in% centered)) {
-      warning("Moderators, outcome variables, and weights/offsets",
-              " cannot be centered.")
+      warn_wrap("Moderators, outcome variables, and weights/offsets
+                cannot be centered.")
       centered <- centered[centered %nin% omitvars]
     }
     if (length(centered) > 0) {
@@ -422,7 +423,7 @@ center_ss_non_survey <- function(d, weights, facvars = NULL, fvars, pred,
     # Dealing with two-level factors that aren't part
     # of an interaction/focal pred
     for (v in fv2) {
-      if (is.factor(d[[v]]) & length(unique(d[[v]])) == 2) {
+      if (is.factor(d[[v]]) && length(unique(d[[v]])) == 2) {
 
         facvars <- c(facvars, v)
 
@@ -431,8 +432,9 @@ center_ss_non_survey <- function(d, weights, facvars = NULL, fvars, pred,
 
   }
 
-  # Fixes a data type error with predict() later
-  d <- as.data.frame(d)
+  if (!is.null(at)) {
+    d <- set_at(at = at, d = d)
+  }
 
   out <- list(d = d, facvars = facvars, fvars = fvars, design = NULL)
 
@@ -444,9 +446,9 @@ center_ss_non_survey <- function(d, weights, facvars = NULL, fvars, pred,
 
 center_ss_survey <- function(d, weights, facvars = NULL, fvars, pred, resp,
                              modx, survey, design, mod2, wname, offname,
-                             centered) {
+                             centered, at) {
 
-  omitvars <- c(pred, resp, modx, mod2, wname, offname)
+  omitvars <- c(pred, resp, modx, mod2, wname, offname, names(at))
 
   # Dealing with two-level factors that aren't part of an interaction
   # /focal pred
@@ -456,8 +458,8 @@ center_ss_survey <- function(d, weights, facvars = NULL, fvars, pred, resp,
   if (centered[1] != "all" && centered[1] != "none") {
 
     if (any(omitvars %in% centered)) {
-      warning("Moderators, outcome variables, and weights/offsets",
-              " cannot be centered.")
+      warn_wrap("Moderators, outcome variables, and weights/offsets cannot be 
+                 centered.")
       centered <- centered[centered %nin% omitvars]
     }
     design <- gscale(vars = centered, data = design, center.only = TRUE)
@@ -475,19 +477,15 @@ center_ss_survey <- function(d, weights, facvars = NULL, fvars, pred, resp,
     }
 
   } else if (centered == "none") {
-
     # Dealing with two-level factors that aren't part
     # of an interaction/focal pred
     for (v in fv2) {
       if (is.factor(d[[v]]) && length(unique(d[[v]])) == 2) {
-
         facvars <- c(facvars, v)
-
       }
     }
 
   } else if (centered == "all") {
-
     # Center all non-focal
     ndfvars <- fvars[fvars %nin% omitvars]
 
@@ -495,13 +493,32 @@ center_ss_survey <- function(d, weights, facvars = NULL, fvars, pred, resp,
       design <- gscale(vars = ndfvars, data = design, center.only = TRUE)
       d <- design$variables
     }
+  }
 
+  if (!is.null(at)) {
+    d <- set_at(at = at, d = d)
   }
 
   out <- list(d = d, design = design, facvars = facvars, fvars = fvars)
 
   return(out)
+}
 
+#### Deal with at variables #################################################
+set_at <- function(at, d) {
+  for (v in names(at)) {
+    if (v %nin% names(d)) stop_wrap("`at` variable ", v, " not found in data.")
+    if (!is.numeric(d[[v]])) {
+      warn_wrap("Inclusion of non-numeric variable ", v, " in `at` argument
+                is not currently supported. As an alternative, treat the 
+                variable as a factor and use the relevel() function to
+                set this value as its reference level before fitting your
+                model.")
+    } else {
+      d[[v]] <- d[[v]] - at[[v]]
+    }
+  }
+  return(d)
 }
 
 #### Send deprecation warnings ##############################################
@@ -922,7 +939,6 @@ drop_factor_levels <- function(d, var, values, labels) {
   return(d)
 
 }
-
 
 # get_contrasts <- function(model) {
 #   form <- as.formula(formula(model))